+++ lib/ui/SearchPanel.dart
@@ -33,6 +33,10 @@ import 'package:sprintf/sprintf.dart';
 import 'athletics/AthleticsGameDetailPanel.dart';
 
 class SearchPanel extends StatefulWidget {
+  final Map<String, dynamic> searchData;
+
+  const SearchPanel({Key key, this.searchData}) : super(key: key);
+
   @override
   _SearchPanelState createState() => _SearchPanelState();
 }
@@ -235,8 +239,16 @@ class _SearchPanelState extends State<SearchPanel> {
           AthleticsGameDetailPanel(gameId: event.speaker, sportName: event.registrationLabel,)));
     }
     else {
+      String groupId = AppJson.stringValue(widget.searchData!= null ? widget.searchData["group_id"] : null);
       Navigator.push(context, CupertinoPageRoute(builder: (context) =>
-          ExploreDetailPanel(explore: explore)));
+          ExploreDetailPanel(explore: explore, browseGroupId: groupId,))).
+            then(
+              (value){
+                if(value!=null && value == true){
+                  Navigator.pop(context, true);
+                }
+              }
+            );
     }
   }